+++ packages/analytics/src/components/ChartAnalytics/models/gauge.ts
@@ -94,7 +94,8 @@ export class DHIS2GaugeChart extends DHIS2Chart {
 		const legendColor = legendSet
 			? getColorFromLegendSet(
 					Array.isArray(legendSet)
-						? head(legendSet).legends
+						? (head(legendSet as unknown as Array<LegendSet>)
+								?.legends ?? [])
 						: (legendSet as LegendSet).legends,
 					this.getValue(),
 				)