+++ packages/analytics/src/components/DHIS2PivotTable/components/TableBody/index.tsx
@@ -5,8 +5,9 @@ import { DataTableCell, DataTableRow, TableBody } from "@dhis2/ui";
 import React, { Fragment, ReactNode, useRef } from "react";
 import { AnalyticsItem } from "@hisptz/dhis2-utils";
 import classes from "./TableBody.module.css";
-import { DHIS2Dimension } from "../../interfaces/index.js";
+import { DHIS2Dimension } from "../../interfaces";
 import { useResizeObserver } from "usehooks-ts";
+import { getTextColorFromBackgroundColor } from "../../utils/color";
 
 function DataRowRenderer({
 	mapper,
@@ -25,17 +26,36 @@ function DataRowRenderer({
 		...mapper,
 		[dimension]: item.uid,
 	}));
+
 	return (
 		<>
-			{completeMapper?.map((mapper) => (
-				<DataTableCell
-					key={`${Object.values(mapper).join("-")}-value`}
-					align="center"
-					bordered
-				>
-					{engine?.getValue(mapper) ?? ""}
-				</DataTableCell>
-			))}
+			{completeMapper?.map((mapper) => {
+				const legend = engine?.getItemValueLegend(mapper);
+
+				return (
+					<DataTableCell
+						style={{
+							background:
+								legend?.style === "FILL"
+									? legend?.color
+									: undefined,
+							color:
+								legend?.style === "FILL"
+									? getTextColorFromBackgroundColor(
+											legend?.color,
+										)
+									: legend?.style === "TEXT"
+										? legend.color
+										: undefined,
+						}}
+						key={`${Object.values(mapper).join("-")}-value`}
+						align="center"
+						bordered
+					>
+						{engine?.getValue(mapper) ?? ""}
+					</DataTableCell>
+				);
+			})}
 		</>
 	);
 }